datacollector_service/requirements.txt
grpcio
grpcio-tools
mysql-connector-python
yfinance
pybreaker
confluent-kafka
prometheus_client